Retrieves organizations of authenticated user
This code authenticates via the parameters provided on the CLI, looks to see who's authenticated (that's the "" part) and then snags all their organizations. Note that the Organizations.List returns a slice of pointers that need to be pointed to or you'll print out gibberish, like I did for a while.