Now retrieving READMEs :+1:
After collecting a list of the organization's repositories, I need to use .Name, not .FullName, since FullName includes the organization and the call to GetReadme also appends the owner, which is the org here. This all resulted in a query to places like `github.com/repos/intelsdi-x/intelsdi-x/snap-plugin-publisher-kafka/` which doesn't exist.

TL;DR - using .Name now and able to retrieve the README.